requirements.txt
pytest
pytest-playwright
pytest-html
pytest-metadata